Move specs close to implementations.
